requirements.txt
accelerate==0.30.0
certifi==2022.12.7
charset-normalizer==2.1.1
diffusers==0.27.1
filelock==3.13.1
fsspec==2024.6.1
hf-xet==1.1.10
huggingface-hub==0.24.2
idna==3.4
imageio==2.37.0
importlib_metadata==8.7.0
Jinja2==3.1.4
lazy_loader==0.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.0
opencv-python==4.8.1.78
packaging==25.0
peft==0.10.0
pillow==11.0.0
pip-chill==1.0.3
psutil==7.1.0
PyYAML==6.0.2
regex==2025.9.18
requests==2.28.1
safetensors==0.6.2
scikit-image==0.22.0
scipy==1.15.3
sympy==1.13.3
tifffile==2025.5.10
tokenizers==0.15.2
torch==2.1.0
tqdm==4.67.1
transformers==4.37.2
triton==2.1.0
typing_extensions==4.12.2
urllib3==1.26.13
xformers==0.0.22.post7
zipp==3.23.0
matplotlib==3.10.6
